python-apps/item-detection/requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
tqdm>=4.41.0
matplotlib>=3.2.0
tensorboard>=2.4.1
seaborn>=0.11.0
scipy>=1.4.1
ultralytics
